+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ Fixed
 - **Projectroles**
     - CSS padding issue with ``sodar-list-btn`` and Chrome (#529, sodar#844)
     - Crash from missing optional setting ``PROJECTROLES_DISABLE_CATEGORIES`` (#524)
+    - Remote project editing not prevented in REST API views (#523)
 
 Removed
 -------